backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
boto3==1.34.108
python-dotenv==1.0.1
requests==2.31.0
PyJWT==2.8.0
bcrypt==4.3.0
passlib==1.7.4
tzdata==2025.3
pytz==2024.1
